Specify toolchain rev and lock dependencies.
This is a library, but in general it isn't safe to name a compiler
revision on nightly without _also_ pinning one's dependencies. Found
this out the hard way when Cargo pulled in a version of some crates that
assumed that "nightly" meant "can use C string literals", but I was
using a nightly compiler too old for that.